refactor(user): user.service, permission.service를 service/ 하위로 이동
schedule, resource 모듈과 동일한 구조로 맞추기 위해 이동
- user.service.ts → service/user.service.ts
- permission.service.ts → service/permission.service.ts
- 참조하는 전체 파일 import 경로 일괄 수정
