+++ src/user/service/user-admin.service.ts
@@ -3,7 +3,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User, UserStatus } from '../user.entity';
 import { UpdateUserInfoDto, UserListFilter, FindFilteredResult } from '../dto/user-admin.dto';
-import { UserService } from '../user.service';
+import { UserService } from './user.service';
 
 @Injectable()
 export class UserAdminService {
